requirements.txt
pytrends

ebaysdk

requests

psycopg2-binary

sqlalchemy

pandas

prophet

xgboost

scikit-learn

langchain

langchain-openai

streamlit

python-dotenv

urllib3

plotly

fastapi 

uvicorn
